Fix TimberHandler so that if the process where the handler is created gets forked, the child process can keep logging. (#15)
In essence, this patch makes the queue in the handler be a multiprocessing.Queue instead of queue.Queue. This change makes all interactions safe to happen both in different processes and threads. Another side effect of dealing with multiple processes is the fact that the Flusher can not/should not run as a thread of each of the childs. The suggested solution only starts the flush thread if we detect the running process is the main one.

Existing tests have been adapted so that they are compatible with the multiprocessing.Queue interface which .qsize() method is not implemented in all platforms.